layout changes for adding the playlist ui
