+++ global.d.ts
@@ -1,2 +1,4 @@
 declare module 'youtube-api';
 declare module 'moniker';
+declare module 'react-semantic-ui-range';
+declare module 'xml2js';
