+++ src/components/ChatPlaylistCard/ChatVideoCard.test.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// TODO: Add testing here with enzyme
+
+describe('rendering', () => {
+  it('renders with title, duration and thumbnail');
+  it('renders without a thumbnail');
+});
