+++ src/index.d.ts
@@ -36,4 +36,6 @@ interface PlaylistVideo {
   url: string;
   name: string;
   img?: string;
+  channel: string;
+  duration: string;
 }
